+++ src/simulate/data.cpp
@@ -254,7 +254,7 @@ namespace CASTLE {
    std::vector<std::vector<unsigned int> > electron_integration_list;
    std::vector<std::vector<unsigned int> > electron_nearest_electron_list;
    std::vector<std::vector<unsigned int> > electron_nearest_atom_list;
-   std::vector<std::vector<unsigned int> > electron_ee_scattering_list;
+   std::vector<std::vector<uint32_t> > electron_ee_scattering_list;
    std::vector<std::vector<unsigned int> > electron_ea_scattering_list;
    std::vector<std::vector< int> > cell_lattice_coordinate;
    std::vector<std::vector<unsigned int > > cell_integration_lists;
@@ -347,7 +347,7 @@ namespace CASTLE {
    void ea_scattering(const int e, const int array_index);
    void ee_scattering();
    int ee_inelastic(const int electron, const int electron_collision, const double deltaE);
-   double ee_elestic(const int electron, const int electron_collision,  const double e_energy, const double d_e_energy);
+   int ee_elestic(const int electron, const int electron_collision,  const double e_energy, const double d_e_energy);
 
    double B_E_distrib(const double& epsilon);
    double M_B_distrib(const double& epsilon, const double& beta);
